+++ ErsatzTV.Application/Configuration/Queries/GetMpegTsScripts.cs
@@ -1,0 +1,5 @@
+using ErsatzTV.Core.FFmpeg;
+
+namespace ErsatzTV.Application.Configuration;
+
+public record GetMpegTsScripts : IRequest<List<MpegTsScript>>;