+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - Add `MediaItem_Resolution` template data (the current `Resolution` variable is the FFmpeg Profile resolution)
   - Add `MediaItem_Start` template data (DateTimeOffset)
   - Add `MediaItem_Stop` template data (DateTimeOffset)
-  - Add `ScaledResolution` (the final size of the frame before padding)
+  - Add `ScaledResolution` template data (the final size of the frame before padding)
   - Add `place_within_source_content` (true/false) field to image graphics element
 - Classic schedules: add collection type `Search Query`
   - This allows defining search queries directly on schedule items without creating smart collections beforehand
@@ -18,6 +18,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
     - Filter: `smart_collection:"sd movies" AND plot:"christmas"`
     - Combine: `smart_collection:"old commercials" OR smart_collection:"nick promos"`
 - Scripted schedules: add `custom_title` to `start_epg_group`
+- Add MPEG-TS Script system
+  - This allows using something other than ffmpeg (e.g. streamlink) to concatenate segments back together when using MPEG-TS streaming mode
+  - Scripts live in config / scripts / mpegts
+  - Each script gets its own subfolder which contains an `mpegts.yml` definition and corresponding windows (powershell) and linux (bash) scripts
 
 ### Fixed
 - Fix HLS Direct playback with Jellyfin 10.11